+++ src/components/dialog/dialog.tsx
@@ -78,16 +78,11 @@ export class Dialog {
     this.handleTransitionEnd = this.handleTransitionEnd.bind(this);
     this.handleKeyDown = this.handleKeyDown.bind(this);
     this.handleOverlayClick = this.handleOverlayClick.bind(this);
-    this.updateSlots = this.updateSlots.bind(this);
+    this.handleSlotChange = this.handleSlotChange.bind(this);
 
     this.modal = new Modal(this.host, {
       onFocusOut: () => this.panel.focus()
     });
-  }
-
-  componentWillLoad() {
-    this.updateSlots();
-    this.host.shadowRoot.addEventListener('slotchange', this.updateSlots);
   }
 
   componentDidLoad() {
@@ -99,8 +94,6 @@ export class Dialog {
 
   disconnectedCallback() {
     unlockBodyScrolling(this.host);
-
-    this.host.shadowRoot.removeEventListener('slotchange', this.updateSlots);
   }
 
   /** Shows the dialog */
@@ -162,6 +155,10 @@ export class Dialog {
     }
   }
 
+  handleSlotChange() {
+    this.hasFooter = hasSlot(this.host, 'footer');
+  }
+
   handleTransitionEnd(event: TransitionEvent) {
     const target = event.target as HTMLElement;
 
@@ -174,10 +171,6 @@ export class Dialog {
         this.panel.focus();
       }
     }
-  }
-
-  updateSlots() {
-    this.hasFooter = hasSlot(this.host, 'footer');
   }
 
   render() {
@@ -222,7 +215,7 @@ export class Dialog {
           </div>
 
           <footer part="footer" class="dialog__footer">
-            <slot name="footer" />
+            <slot name="footer" onSlotchange={this.handleSlotChange} />
           </footer>
         </div>
       </div>